requirements.txt
# Core libraries
streamlit>=1.28
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
pillow

# Jupyter (optional for notebook debugging)
jupyter
ipykernel

# Inference-related
tensorflow  # or torch, depending on your model